BH17XX_Driver
User just call the begin function from the application by passing mode and sensor type in the begin function then in the begin it will set values according to given sensor and resolution mode and then also settle down the values for power on and power down Commands and then check the status of device and return a function of initlization function where first send a command for reset and then set measuremnt time of light sensor and then inilize the sensor by sending the command of resolution setted by the user and read the light level this work in the dfault case.
user can also set the resoltion mode and timing mode by setter and getter functions by and also send command in different timings and resolution mode and also set resoltuin factor and measurement time accoding to his will.